Fixed loose type cast in 64bit architecture

+++ MTFontIcon/NSString+Unicode.m
@@ -21,8 +21,8 @@
     NSUInteger intValue = 0;
     for (int i = 0; i < length; i++) {
         char c = [[self lowercaseString] characterAtIndex:i];
-        int number = (int)c;
-        int realNumber = 0;
+        NSUInteger number = (NSUInteger)c;
+        NSUInteger realNumber = 0;
         NSUInteger numericShift = 48;
         NSUInteger charShift = 97;
